+++ CHANGELOG.md
@@ -3,6 +3,8 @@ next-version
 
 - cli: creating static html reports is no-longer supported.
 - http: add support for LOGREDUCE_CA_EXTRA or /etc/pki/tls/certs/ca-extra.crt to add an extra ca to the default bundle.
+- api: provides a /ready and /metrics endpoint for monitoring.
+- web: minor improvement to the user interface.
 
 0.9.1
 =====